+++ CPP_Files/main.cpp
@@ -1,0 +1,16 @@
+//main
+
+//#include <iostream>
+//#include <string>
+#include "Calendar.hpp"
+
+using namespace std;
+
+int main()
+{
+	Calendar cal;
+	cal.readInput();
+	cal.writeOutput();
+ 
+	return 0;
+}
