--- formFunc.hpp
@@ -1,17 +1,0 @@
-//non-member form functions headers
-#ifndef FORMFUNC_HPP
-#define FORMFUNC_HPP
-
-#include <iostream>
-#include <string>
-
-using namespace std;
-
-int readDate(string s);
-int readTime(string s);
-string readName(string s);
-string readTags(string s);
-string timeCat(int t);
-string dateCat(int d);
-
-#endif
