--- main.cpp
@@ -1,16 +1,0 @@
-//main
-
-//#include <iostream>
-//#include <string>
-#include "Calendar.hpp"
-
-using namespace std;
-
-int main()
-{
-	Calendar cal;
-	cal.readInput();
-	cal.writeOutput();
- 
-	return 0;
-}
